pyproject.toml
[tool.poetry]
name = "speech-defect-detector"
version = "0.1.0"
description = "Detection of speech defects using tongue twisters"
authors = ["Kuzmin Dmitry Andreevich"]
readme = "README.md"
packages = [{include = "speech_defect_detector"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.0"
pytorch-lightning = "^2.1.0"
torchmetrics = "^1.2.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
mlflow = "^2.9.0"
dvc = "^3.38.0"
dvc-s3 = "^3.0.0"
fire = "^0.5.0"
numpy = "^1.24.0"
librosa = "^0.10.1"
scikit-learn = "^1.3.0"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
isort = "^5.12.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "W191"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

